Add files via upload
A new version of the JobChainGen. It only works with CCL and SH-files and therefore almost no data is generated by the script. Also it gives more possibilities for simulations (Iterations and Timesteps can be altered through the JobChain)

As this was rewritten when i was ill and lazy, not all functions are fully functional. So a call this a BETA-version.